pyproject.toml
[tool.poetry]
name = "intents"
version = "0.0.1"
description = "Define and operate Dialogflow Agents with a simple, code-first, approach"
authors = ["Dario <dario.chi@inventati.org>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.4.1"
google-cloud-dialogflow = "^2.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
ipython = "^7.19.0"
notebook = "^6.2.0"
ipykernel = "^5.4.3"
sphinx = "^3.5.1"
recommonmark = "^0.7.1"
sphinx-rtd-theme = "^0.5.1"
autopep8 = "^1.5.6"
m2r2 = "^0.2.7"
sphinx-markdown-tables = "^0.0.15"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
